pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown-link-prefix"
version = "0.1.0"
description = "Python-Markdown extension: expand prefixed link shorthand (e.g. youtube:VIDEO_ID, snkb:KB123, snci:sys_id) to full URLs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [{ name = "BarCar" }]
keywords = ["markdown", "extension", "links", "youtube", "servicenow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "markdown>=3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-cov",
    "pyyaml>=6.0",
]
docs = [
    "zensical",
    "pymdown-extensions",
]
yaml = [
    "pyyaml>=6.0",
]

[project.entry-points."markdown.extensions"]
link_prefix = "markdown_link_prefix:LinkPrefixExtension"

[project.urls]
Documentation = "https://barcar.github.io/markdown-link-prefix/"
Repository = "https://github.com/barcar/markdown-link-prefix"

[tool.setuptools.packages.find]
where = ["."]
include = ["markdown_link_prefix*"]
